final submission/Cyber-Risk-and-Threat-Intelligence-platform-main (1)/Cyber-Risk-and-Threat-Intelligence-platform-main/requirements.txt
python-dotenv==1.2.2
Requests==2.32.5
streamlit
pandas
plotly
